1、将所有金额改成Integer类型，使用分作为单位（同时修改数据库中对应金额字段类型） 2、优化APP支付返回对象，并将前台需要的签名数据在后台完成后在返回 3、优化获取聚合支付的返回对象 4、优化表字段将sysno修改为sys_no
待开发：
1、分账处理逻辑
 (1)非分账支付的支付订单无法调用分账接口
 (2)只有完成支付的订单可以调用分账接口
 (3)如果已经退款则无法调用分账
2、分账记录数据表
3、分账回退数据表
4、还有部分已xxx_sysno命名的字段需要修改

+++ pay/src/main/java/com/yp/pay/entity/aliandwx/req/AliF2FPayReq.java
@@ -101,6 +101,6 @@ public class AliF2FPayReq {
      * 但传入了【订单总金额】和【可打折金额】，则该值默认为【订单总金额】-【可打折金额】
      */
     @ApiModelProperty(value = "不参与优惠计算的金额", hidden = true)
-    private BigDecimal undiscountableAmount;
+    private Integer undiscountableAmount;
 
 }
